move some models and clean up a little
