+++ joplin/base/models/__init__.py
@@ -21,26 +21,13 @@ from wagtail.admin.edit_handlers import PageChooserPanel
 from base import blocks as custom_blocks
 from base import forms as custom_forms
 
+from .janis_page import JanisPage, JanisBasePage
+from .service_page import ServicePage
+from .translated_image import TranslatedImage
+
 WYSIWYG_GENERAL = ['h1', 'h2', 'h3', 'h4', 'bold', 'link', 'ul', 'ol', 'code']
-WYSIWYG_SERVICE_STEP = ['ul', 'ol', 'link', 'code']
 DEFAULT_MAX_LENGTH = 255
 SHORT_DESCRIPTION_LENGTH = 300
-
-class TranslatedImage(AbstractImage):
-    admin_form_fields = Image.admin_form_fields
-
-    def __str__(self):
-        return self.title or self.title_en
-
-
-class TranslatedImageRendition(AbstractRendition):
-    image = models.ForeignKey(TranslatedImage, related_name='renditions', on_delete=models.CASCADE)
-
-    class Meta:
-        unique_together = (
-            ('image', 'filter_spec', 'focal_point_key'),
-        )
-
 
 @register_snippet
 class ThreeOneOne(ClusterableModel):
@@ -50,201 +37,12 @@ class ThreeOneOne(ClusterableModel):
     def __str__(self):
         return self.title
 
-
 class HomePage(Page):
     parent_page_types = []
     # subpage_types = ['base.ServicePage', 'base.ProcessPage', 'base.InformationPage', 'base.DepartmentPage']
     subpage_types = ['base.ServicePage', 'base.ProcessPage', 'base.InformationPage', 'base.DepartmentPage', 'base.TopicPage']
 
     image = models.ForeignKey(TranslatedImage, null=True, on_delete=models.SET_NULL, related_name='+')
-
-
-class JanisBasePage(Page):
-    parent_page_types = ['base.HomePage']
-    subpage_types = []
-    search_fields = Page.search_fields + [
-        index.RelatedFields('owner', [
-            index.SearchField('last_name', partial_match=True),
-            index.FilterField('last_name'),
-        ])
-    ]
-
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    author_notes = RichTextField(
-        # max_length=DEFAULT_MAX_LENGTH,
-        features=['ul', 'ol', 'link'],
-        blank=True,
-        verbose_name='Notes for authors (Not visible on the resident facing site)'
-    )
-
-    def janis_url(self):
-        page_slug = self.slug
-
-        if self.janis_url_page_type == "department":
-            return os.environ["JANIS_URL"] + "/en/" + page_slug
-
-        if self.janis_url_page_type == "topiccollection":
-            theme_slug = self.theme.slug;
-            return os.environ["JANIS_URL"] + "/en/" + theme_slug + "/" + page_slug
-
-        if self.janis_url_page_type == "topic":
-            # If we have a topic collection
-            if self.topiccollections and self.topiccollections.all():
-                theme_slug = self.topiccollections.all()[0].topiccollection.theme.slug;
-                tc_slug = self.topiccollections.all()[0].topiccollection.slug;
-                return os.environ["JANIS_URL"] + "/en/" + theme_slug + "/" + tc_slug + "/" + page_slug            
-
-
-        if self.janis_url_page_type == "services" or self.janis_url_page_type == "information":
-            # If we have topics, use the first one
-            if self.topics and self.topics.all():
-                topic_slug = self.topics.all()[0].topic.slug
-                # Make sure we have a topic collection too
-                if self.topics.all()[0].topic.topiccollections.all():
-                    theme_slug = self.topics.all()[0].topic.topiccollections.all()[0].topiccollection.theme.slug;
-                    tc_slug = self.topics.all()[0].topic.topiccollections.all()[0].topiccollection.slug;
-                    return os.environ["JANIS_URL"] + "/en/" + theme_slug + "/" + tc_slug + "/" + topic_slug + "/" + page_slug
-
-            # TODO: bring back departments now that we can have multiple
-
-        # We don't have a valid live url
-        # TODO: add something to make this clear to users
-        return "#"
-
-    def janis_preview_url(self):
-        revision = self.get_latest_revision()
-        url_page_type = self.janis_url_page_type
-        global_id = graphene.Node.to_global_id('PageRevisionNode', revision.id)
-
-        return os.environ["JANIS_URL"] + "/en/preview/" + url_page_type + "/" + global_id
-
-    # Default preview_url before janis_preview_url gets set
-    def fallback_preview_url(self):
-        return "https://alpha.austin.gov"
-
-    # data needed to construct preview URLs for any language
-    # [janis_url_base]/[lang]/preview/[url_page_type]/[global_id]
-    # ex: http://localhost:3000/es/preview/information/UGFnZVJldmlzaW9uTm9kZToyMjg=
-    def preview_url_data(self):
-        revision = self.get_latest_revision()
-        global_id = graphene.Node.to_global_id('PageRevisionNode', revision.id)
-
-        return {
-            "janis_url_base": os.environ["JANIS_URL"],
-            "url_page_type": self.janis_url_page_type,
-            "global_id": global_id
-        }
-
-    class Meta:
-        abstract = True
-
-class JanisPage(JanisBasePage):
-    @cached_classmethod
-    def get_edit_handler(cls):
-        if hasattr(cls, 'edit_handler'):
-            return cls.edit_handler.bind_to_model(cls)
-
-        edit_handler = TabbedInterface([
-            ObjectList(cls.content_panels + [
-                FieldPanel('author_notes')
-            ], heading='Content'),
-            ObjectList(Page.promote_panels + cls.promote_panels, heading='Search Info')
-        ])
-
-        return edit_handler.bind_to_model(cls)
-
-    class Meta:
-        abstract = True
-
-class ServicePage(JanisPage):
-    janis_url_page_type = "services"
-
-    steps = StreamField(
-        [
-            ('basic_step', RichTextBlock(
-                features=WYSIWYG_SERVICE_STEP,
-                label='Basic Step'
-            )),
-            ('step_with_options_accordian', StructBlock(
-                [
-                    ('options_description', TextBlock('Describe the set of options')),
-                    ('options', ListBlock(
-                        StructBlock([
-                            ('option_name', TextBlock(
-                                label='Option name. (When clicked, this name will expand the content for this option'
-                            )),
-                            ('option_description', RichTextBlock(
-                                features=WYSIWYG_SERVICE_STEP,
-                                label='Option Content',
-                            )),
-                        ]),
-                    )),
-                ],
-                label="Step With Options"
-            )),
-        ],
-        verbose_name='Write out the steps a resident needs to take to use the service',
-        # this gets called in the help panel
-        help_text='A step may have a basic text step or an options accordion which reveals two or more options',
-        blank=True
-    )
-
-    dynamic_content = StreamField(
-        [
-            ('map_block', custom_blocks.SnippetChooserBlockWithAPIGoodness('base.Map', icon='site')),
-            ('what_do_i_do_with_block', custom_blocks.WhatDoIDoWithBlock()),
-            ('collection_schedule_block', custom_blocks.CollectionScheduleBlock()),
-            ('recollect_block', custom_blocks.RecollectBlock()),
-        ],
-        verbose_name='Add any maps or apps that will help the resident use the service',
-        blank=True
-    )
-    additional_content = RichTextField(
-        features=WYSIWYG_GENERAL,
-        verbose_name='Write any additional content describing the service',
-        help_text='Section header: What else do I need to know?',
-        blank=True
-    )
-
-    base_form_class = custom_forms.ServicePageForm
-
-    short_description = models.TextField(
-        max_length=SHORT_DESCRIPTION_LENGTH,
-        blank=True,
-        verbose_name='Write a description of this service'
-    )
-
-    content_panels = [
-        FieldPanel('title_en'),
-        FieldPanel('title_es'),
-        FieldPanel('title_ar'),
-        FieldPanel('title_vi'),
-        FieldPanel('short_description'),
-        InlinePanel('topics', label='Topics'),
-        InlinePanel('related_departments', label='Related Departments'),
-        MultiFieldPanel(
-        [
-            HelpPanel(steps.help_text, classname="coa-helpPanel"),
-            StreamFieldPanel('steps')
-        ],
-        heading=steps.verbose_name,
-        classname='coa-multiField-nopadding'
-        ),
-        StreamFieldPanel('dynamic_content'),
-        MultiFieldPanel(
-        [
-            HelpPanel(additional_content.help_text, classname="coa-helpPanel"),
-            FieldPanel('additional_content')
-        ],
-        heading=additional_content.verbose_name,
-        classname='coa-multiField-nopadding'
-        )
-        ,
-        InlinePanel('contacts', label='Contacts'),
-    ]
-
 
 class ProcessPage(JanisPage):
     janis_url_page_type = "processes"
