+++ joplin/base/models/constants.py
@@ -1,0 +1,2 @@
+WYSIWYG_GENERAL = ['h1', 'h2', 'h3', 'h4', 'bold', 'link', 'ul', 'ol', 'code']
+SHORT_DESCRIPTION_LENGTH = 300
